requirements.txt
asgiref==3.5.0
Django==3.1.7
django-tastypie==0.14.4
PyMySQL==1.0.2
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
